01-08-16

+++ subjectmanagement/process_approve_subject.php
@@ -11,7 +11,7 @@
 			
 		for($i=0; $i<$rowCount; $i++)
 		{
-			$result = mysql_query("SELECT * FROM  fr_stud_subject WHERE subject_id = '".$_POST['subject'][$i]."'");
+			$result = mysql_query("SELECT * FROM  fr_stud_subject WHERE subject_id = '".$_POST['subject'][$i]."' AND user_id ='".$_POST['user_id'][$i]."'");
 			if(mysql_num_rows($result) > 0)
 			{
 				$row = mysql_fetch_array($result);
@@ -20,7 +20,7 @@
 				$time = date ("H:i:s");
 
 
-				mysql_query("UPDATE fr_stud_subject SET Date_Created = '".$date."', Time_Created = '".$time."', status = 'APPROVED' WHERE subject_id = '".$_POST['subject'][$i]."'");
+				mysql_query("UPDATE fr_stud_subject SET Date_Created = '".$date."', Time_Created = '".$time."', status = 'APPROVED' WHERE subject_id = '".$_POST['subject'][$i]."' AND user_id ='".$_POST['user_id'][$i]."'");
 
 				if(!file_exists(".".$path))
     			{
@@ -44,7 +44,7 @@
 		$rowCount = count($_POST['subject']);
 		for($i=0; $i<$rowCount; $i++)
 		{
-			mysql_query("DELETE FROM fr_stud_subject WHERE subject_id = '".$_POST['subject'][$i]."'");
+			mysql_query("DELETE FROM fr_stud_subject WHERE subject_id = '".$_POST['subject'][$i]."' AND user_id ='".$_POST['user_id'][$i]."'");
 		}
 		echo '<script> alert("Successfully DISAPPROVE"); window.location.href="../subjectmanagement.php?subject=approve"; </script>';
